+++ Products/PluggableAuthService/plugins/ZODBGroupManager.py
@@ -26,6 +26,8 @@ from zope.event import notify
 from zope.interface import Interface
 
 from Products.PluggableAuthService.events import GroupCreated
+from Products.PluggableAuthService.events import PrincipalAddedToGroup
+from Products.PluggableAuthService.events import PrincipalRemovedFromGroup
 from Products.PluggableAuthService.interfaces.plugins import \
     IGroupEnumerationPlugin
 from Products.PluggableAuthService.interfaces.plugins import IGroupsPlugin
@@ -273,6 +275,7 @@ class ZODBGroupManager(BasePlugin):
         if not already:
             new = current + (group_id,)
             self._principal_groups[principal_id] = new
+            notify(PrincipalAddedToGroup(principal_id, group_id))
             self._invalidatePrincipalCache(principal_id)
 
         return not already
@@ -299,6 +302,7 @@ class ZODBGroupManager(BasePlugin):
 
         if already:
             self._principal_groups[principal_id] = new
+            notify(PrincipalRemovedFromGroup(principal_id, group_id))
             self._invalidatePrincipalCache(principal_id)
 
         return already